+++ string/nlibstring.cpp
@@ -5,7 +5,7 @@ namespace nlib{
 	namespace string{
 
 		String::String(const char *ptr){
-			unsigned int index = 0;
+			size_t index = 0;
 			while(ptr[index] != '\0'){
 				int length = nlib::uchar::UChar::headCount(ptr[index]);
 				if(length == 0)
@@ -15,7 +15,7 @@ namespace nlib{
 			}
 		}
 
-		unsigned int String::size(){
+		size_t String::size(){
 			return buffer.size();
 		}
